requirements.txt
flask-script
jinja2
markupsafe
werkzeug
itsdangerous
coverage
fake-factory
flask
flask-admin
flask-babel
flask-bower
flask-classy
flask-debugtoolbar
flask-login
flask-mail
flask-migrate
flask-openid
flask-sqlalchemy
flask-whooshalchemy
flask-wtf
flipflop
guess_language
gunicorn
psycopg2
sqlalchemy-migrate
wtforms-components
# Optional
ipython
pudb
